+++ packages/styled-react/src/components/PageHeader.tsx
@@ -75,18 +75,35 @@ const PageHeaderTitleArea: ForwardRefComponent<'div', PageHeaderTitleAreaProps> 
   ${sx}
 `
 
-const PageHeader = Object.assign(PageHeaderImpl, {
+type PageHeaderComponent = ForwardRefComponent<'div', PageHeaderProps> & {
+  Actions: typeof PageHeaderActions
+  ContextArea: typeof PrimerPageHeader.ContextArea
+  ParentLink: typeof PrimerPageHeader.ParentLink
+  ContextBar: typeof PrimerPageHeader.ContextBar
+  TitleArea: typeof PageHeaderTitleArea
+  ContextAreaActions: typeof PrimerPageHeader.ContextAreaActions
+  LeadingAction: typeof PrimerPageHeader.LeadingAction
+  Breadcrumbs: typeof PrimerPageHeader.Breadcrumbs
+  LeadingVisual: typeof PrimerPageHeader.LeadingVisual
+  Title: typeof PageHeaderTitle
+  TrailingVisual: typeof PrimerPageHeader.TrailingVisual
+  Description: typeof PrimerPageHeader.Description
+  TrailingAction: typeof PrimerPageHeader.TrailingAction
+}
+
+const PageHeader: PageHeaderComponent = Object.assign(PageHeaderImpl, {
   Actions: PageHeaderActions,
-  Title: PageHeaderTitle,
-  TitleArea: PageHeaderTitleArea,
   ContextArea: PrimerPageHeader.ContextArea,
-  ContextAreaActions: PrimerPageHeader.ContextAreaActions,
-  TrailingVisual: PrimerPageHeader.TrailingVisual,
-  Description: PrimerPageHeader.Description,
+  ParentLink: PrimerPageHeader.ParentLink,
   ContextBar: PrimerPageHeader.ContextBar,
+  TitleArea: PageHeaderTitleArea,
+  ContextAreaActions: PrimerPageHeader.ContextAreaActions,
   LeadingAction: PrimerPageHeader.LeadingAction,
   Breadcrumbs: PrimerPageHeader.Breadcrumbs,
   LeadingVisual: PrimerPageHeader.LeadingVisual,
+  Title: PageHeaderTitle,
+  TrailingVisual: PrimerPageHeader.TrailingVisual,
+  Description: PrimerPageHeader.Description,
   TrailingAction: PrimerPageHeader.TrailingAction,
 })
