+++ packages/styled-react/src/components/RadioGroup.tsx
@@ -1,0 +1,45 @@
+import {Box, RadioGroup as PrimerRadioGroup, type RadioGroupProps as PrimerRadioGroupProps} from '@primer/react'
+import React, {type PropsWithChildren} from 'react'
+import type {SxProp} from '../sx'
+
+export type RadioGroupProps = PropsWithChildren<PrimerRadioGroupProps> & SxProp
+
+const RadioGroupImpl = (props: RadioGroupProps) => {
+  return <Box as={PrimerRadioGroup} {...props} />
+}
+
+// Define local types based on the internal component props
+type CheckboxOrRadioGroupLabelProps = PropsWithChildren<
+  {
+    className?: string
+    visuallyHidden?: boolean
+  } & SxProp
+>
+const CheckboxOrRadioGroupLabel = (props: CheckboxOrRadioGroupLabelProps) => {
+  return <Box as={PrimerRadioGroup.Label} {...props} />
+}
+
+type CheckboxOrRadioGroupCaptionProps = PropsWithChildren<
+  {
+    className?: string
+  } & SxProp
+>
+const CheckboxOrRadioGroupCaption = (props: CheckboxOrRadioGroupCaptionProps) => {
+  return <Box as={PrimerRadioGroup.Caption} {...props} />
+}
+
+type CheckboxOrRadioGroupValidationProps = PropsWithChildren<
+  {
+    className?: string
+    variant: 'error' | 'success'
+  } & SxProp
+>
+const CheckboxOrRadioGroupValidation = (props: CheckboxOrRadioGroupValidationProps) => {
+  return <Box as={PrimerRadioGroup.Validation} {...props} />
+}
+
+export const RadioGroup = Object.assign(RadioGroupImpl, {
+  Label: CheckboxOrRadioGroupLabel,
+  Caption: CheckboxOrRadioGroupCaption,
+  Validation: CheckboxOrRadioGroupValidation,
+})
